generate_code/ftf_repro/requirements.txt
# Core numerical + data
numpy>=1.26
pandas>=2.1
scipy>=1.11

# Stats / HAC regression (recommended)
statsmodels>=0.14

# Configs / serialization
pyyaml>=6.0

# Parquet IO
pyarrow>=14

# Reporting / plotting
matplotlib>=3.8
seaborn>=0.13

# Progress bars
tqdm>=4.66

# Testing
pytest>=7.4

# Optional acceleration
numba>=0.59
